[BREAKING][rollout, diffusion] feat: make pixel responses always uint8
Quantize decoded pixel outputs once at the rollout boundary and expose a fixed uint8 [0, 255] contract throughout trainer, reward, logging, and dump paths. Remove the unshipped response_transport_dtype option and redundant downstream dtype detection while keeping latent outputs in floating point.

This intentionally changes both custom-reward inputs and built-in rewards that previously consumed unquantized floating-point responses.

AI assistance (OpenAI Codex) was used for this change.

+++ tests/utils/reward_score/test_http_scorer_client_on_cpu.py
@@ -64,7 +64,7 @@ async def _run_retry_e2e():
     server_url = f"http://127.0.0.1:{sock.getsockname()[1]}/score"
 
     kwargs = {
-        "solution_image": torch.zeros(3, 2, 2),
+        "solution_image": torch.zeros(3, 2, 2, dtype=torch.uint8),
         "ground_truth": "prompt",
         "server_url": server_url,
         "max_retries": 2,
